DRAFT: add functionality to Ontology (#95)
* WIP - json and base64 encoders

* WIP - implementing consensus protocol

* cast buffer to string for JSON decoding for now

* WIP - implemented compression task

* fixed linux build

* adding Type checking update (#92)

* completed compression implementation

* removed casting to uint16 on tx_count

* add functionality to ontology
fix an issue in extensions when parsing block - wrong tx_hash calculation

* remove print messages

* fix wrong sizing in consensus block in extensions
fix more unit tests

* merge from develop

* fix few issues in extensions

* need to reverse the hash in the extensions

* merge errors

* remove print messages

* remove print messages

+++ lib/libtpe/tpe/task/consensus_ont_block_compression_task.h
@@ -48,7 +48,7 @@ public:
     PSha256_t block_hash();
     PSha256_t compressed_block_hash();
 
-    size_t txn_count();
+    uint32_t txn_count();
 
     const std::vector<unsigned int>& short_ids();
